requirements.txt
django==1.11
django-extensions
django-environ
django-braces
django-suit==0.2.26
psycopg2-binary
pdbpp
yapf
gunicorn
